Ignore changes to resources managed by other operator instances (#63)
* Ignore changes to resources managed by other operator instances

Now that we are running multiple instance of the operator in a cluster, we
need to filter out change notifications for resources that are managed by
other instances.

* Scope feedback controller to specific namespace

The feedback controller was receiving notifications for ClusterOrder
resources in other namespaces. Correct this  by adding namespace scoping
using the same CLOUDKIT_CLUSTER_ORDER_NAMESPACE environment variable as the
main controller.